feat: support different key and auth
