+++ cmd/gocq/main.go
@@ -164,31 +164,27 @@ func LoginInteract() {
 		}
 	}
 	signServer, _ := GetAvaliableSignServer()
-	if signServer != "-" && signServer != "" {
-		log.Infof("使用服务器 %s 进行数据包签名", signServer)
-		if base.SignServerBearer != "-" && base.SignServerBearer != "" {
-			log.Infof("使用 Bearer %s 认证签名服务器 %s ", base.SignServerBearer, signServer)
-		}
+	if len(signServer.URL) > 1 {
 		// 等待签名服务器直到连接成功
 		if !signWaitServer() {
 			log.Fatalf("连接签名服务器失败")
 		}
-		signRegister(base.Account.Uin, device.AndroidId, device.Guid, device.QImei36, base.Key)
+		signRegister(base.Account.Uin, device.AndroidId, device.Guid, device.QImei36, signServer.Key)
 		go signStartRefreshToken(base.Account.RefreshInterval) // 定时刷新 token
 		wrapper.DandelionEnergy = energy
 		wrapper.FekitGetSign = sign
 		if !base.IsBelow110 {
 			if !base.Account.AutoRegister {
-				log.Warn("自动注册实例已关闭，若未配置 sign-server 端自动注册实例则实例丢失时需要重启 go-cqhttp 以正常签名")
+				log.Warn("自动注册实例已关闭，请配置 sign-server 端自动注册实例以保持正常签名")
 			}
 			if !base.Account.AutoRefreshToken {
-				log.Warn("自动刷新 token 已关闭，token 过期后获取签名时将不会立即尝试刷新获取新 token")
+				log.Info("自动刷新 token 已关闭，token 过期后获取签名时将不会立即尝试刷新获取新 token")
 			}
 		} else {
 			log.Warn("签名服务器版本 <= 1.1.0 ，无法使用刷新 token 等操作，建议使用 1.1.6 版本及以上签名服务器")
 		}
 	} else {
-		log.Warnf("警告: 未配置签名服务器, 这可能会导致登录 45 错误码或发送消息被风控")
+		log.Warnf("警告: 未配置签名服务器或签名服务器不可用, 这可能会导致登录 45 错误码或发送消息被风控")
 	}
 
 	if base.Account.Encrypt {
